requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleak==0.22.2
blinker==1.7.0
branca==0.7.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.3.2
debugpy==1.8.2
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
emailmessage==0.2.1
executing==2.0.1
faiss-cpu==1.11.0.post1
fastapi==0.115.11
filelock==3.13.4
filetype==1.2.0
Flask==3.0.3
flask-cors==5.0.1
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gpt-index==0.8.42
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.7
imbalanced-learn==0.12.3
imblearn==0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.1.2
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.3
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.5
langchain==0.2.10
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langsmith==0.1.93
leaflet==0.0.3
libclang==18.1.1
llama-cloud==0.0.9
llama-index==0.10.57
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.12
llama-index-core==0.10.57
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llvmlite==0.44.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.61.0
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
openai==1.37.0
openai-whisper==20240930
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydotplus==2.0.2
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.26.3
pyparsing==3.1.2
pypdf==4.3.1
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.54
rpds-py==0.19.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.0
scipy==1.14.0
seaborn==0.13.2
semantic-text-splitter==0.27.0
sentence-transformers==5.0.0
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.46.0
streamlit==1.37.0
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.4.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.49.0
twilio==9.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.77
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==1.26.19
uvicorn==0.34.0
virtualenv==20.25.1
watchdog==4.0.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
wheel==0.43.0
whisper==1.1.10
winrt-runtime==2.1.0
winrt-Windows.Devices.Bluetooth==2.1.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.1.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.1.0
winrt-Windows.Devices.Enumeration==2.1.0
winrt-Windows.Foundation==2.1.0
winrt-Windows.Foundation.Collections==2.1.0
winrt-Windows.Storage.Streams==2.1.0
wrapt==1.16.0
xxhash==3.5.0
xyzservices==2024.6.0
yarl==1.9.4
